test: split compatibility integration tests into customer-layer only
Restructure Microsoft.DotNet.ApiCompat.IntegrationTests so it exercises
only the customer-facing layer (CLI invocation and MSBuild-target
invocation against a real test project) and runs on every PR under both
.NET Core MSBuild and desktop MSBuild.

* Multi-target the IntegrationTests project to
  $(SdkTargetFramework);$(NetFrameworkToolCurrent) so the net472
  test runtime exercises the Runtime="NET" taskhost path.
* Drop the netcore-only ProjectReferences (PackageValidation,
  ApiCompat.Task as compile reference) and the
  Microsoft.CodeAnalysis.CSharp PackageReference. Add
  ReferenceOutputAssembly="false" build-ordering deps on
  ApiCompat.Task and ApiCompat.Tool with
  SetTargetFramework=TargetFramework=$(NetMinimum) so the netcore
  build of the netcore-only projects is consumed regardless of the
  inner build.
* Move the in-process validator tests to
  Microsoft.DotNet.PackageValidation.Tests:
  - CompatibleFrameworkInPackageValidatorIntegrationTests.cs ->
    CompatibleFrameworkInPackageValidatorTests.cs
  - The five in-process methods of ValidatePackageTargetIntegrationTests
    (ValidatePackageWithReferences, ValidateOnlyErrorWhenAReferenceIsRequired,
    ValidateErrorWhenTypeForwardingReferences,
    ValidateMissingReferencesIsOnlyLoggedWhenRunningWithReferences,
    ValidateReferencesAreRespectedForPlatformSpecificTFMs) ->
    ValidatePackageInProcessTests.cs
* Trim ValidatePackageTargetIntegrationTests.cs to keep only the
  pack-the-asset/assert-stdout tests that exercise the SDK's
  EnablePackageValidation pipeline end-to-end.
* Drop the dead CopyApiCompatFilesForTests target.
* Update test/UnitTests.proj: drop the
  Condition="'$(TestFullMSBuild)' == 'true'" gate on the
  IntegrationTests entry so it runs on every PR with both the default
  $(SdkTargetFramework) row and an unconditional
  TargetFramework=net472/RuntimeTargetFramework=net472 row that drives
  desktop MSBuild. Pre-list the soon-to-be-added GenAPI.IntegrationTests
  in the same desktop-MSBuild block.
